% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_colors_function.R
\name{butterfly_palettes}
\alias{butterfly_palettes}
\title{Select butterfly palette}
\usage{
butterfly_palettes(specie)
}
\arguments{
\item{specie}{Butterfly specie}
}
\value{
A vector of hex colour codes
}
\description{
This function returns colour palettes
}
