% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_colors_function.R
\docType{data}
\name{butterfly_species}
\alias{butterfly_species}
\title{List of butterfly species}
\format{
An object of class \code{list} of length 30.
}
\usage{
butterfly_species
}
\description{
List of butterfly species
}
\keyword{datasets}
